MCR-2359 move to annotation package
